React re-write + Exercises & Progress

+++ api/src/services/PropertyService.ts
@@ -1,5 +1,6 @@
-import { v1 as uuid } from 'uuid';
 import { forwardRef, Inject, Injectable } from '@nestjs/common';
+import { DateTime } from 'luxon';
+import { Context } from 'models/Context';
 import { LanguageId } from 'models/Language';
 import {
   BaseProperty,
@@ -11,16 +12,18 @@ import {
   PropertyType,
   TextProperty,
 } from 'models/Property';
-import { PropertyRepository } from 'repositories/PropertyRepository';
-import { LanguageService } from './LanguageService';
 import { PartOfSpeech } from 'models/Word';
-import * as arrays from 'utils/arrays';
-import { DateTime } from 'luxon';
-import { ChangeService } from './ChangeService';
-import { ChangeBuilder } from './ChangeBuilder';
-import { copy } from 'utils/copy';
-import { Context } from 'models/Context';
 import { DbConnectionManager } from 'repositories/DbConnectionManager';
+import { PropertyRepository } from 'repositories/PropertyRepository';
+import * as arrays from 'utils/arrays';
+import { copy } from 'utils/copy';
+import { v1 as uuid } from 'uuid';
+import { ChangeBuilder } from './ChangeBuilder';
+import { ChangeService } from './ChangeService';
+import { LanguageService } from './LanguageService';
+import { WordService } from './WordService';
+
+const PROPERTY_DELETION_WORDS_THRESHOLD = 50;
 
 export interface CreateBasePropertyParams {
   id?: PropertyId;
@@ -66,6 +69,8 @@ export class PropertyService {
     private propertyRepository: PropertyRepository,
     @Inject(forwardRef(() => LanguageService))
     private languageService: LanguageService,
+    @Inject(forwardRef(() => WordService))
+    private wordService: WordService,
     @Inject(forwardRef(() => ChangeService))
     private changeService: ChangeService,
     private changeBuilder: ChangeBuilder,
@@ -217,6 +222,15 @@ export class PropertyService {
 
   async delete(ctx: Context, { id }: DeletePropertyParams): Promise<Property> {
     const property = await this.propertyRepository.getById(id);
+
+    const wordCount = await this.wordService.getCountByProperty(
+      property.languageId,
+      property.id,
+      property.partOfSpeech,
+    );
+    if (wordCount > PROPERTY_DELETION_WORDS_THRESHOLD) {
+      throw new Error('Property has too many words');
+    }
 
     await this.connectionManager.transactionally(async () => {
       await this.propertyRepository.delete(id);
